:truck: chore(wohoox): remove default export
chore: remove default export createStore
chore: types define move to types
