+++ packages/wohoox/src/index.ts
@@ -1,12 +1,10 @@
-import { createStore } from './core/store'
-
-export default createStore
-export { combineStores } from './core/store'
+export { createStore, combineStores } from './core/store'
 
 export { observer } from './core/observer'
 export { storeMap, defaultStoreName } from './global'
 
 export { isIgnoreCacheKey } from './core/keyCaches'
 
-export type { WohooxPlugin } from './core/plugin'
 export { ignoreToRecordEvent } from './core/plugin'
+
+export * from './types'
