+++ packages/wohoox/src/types.ts
@@ -1,3 +1,5 @@
+export type Options = { strictMode?: boolean; proxySetDeep?: boolean }
+
 export type ActionsDefine<S> = {
   [key: string]: ((state: S, ...args: any) => any) | ActionsDefine<S>
 }
@@ -28,3 +30,18 @@ export type ExtractStores<Stores, Name> = Stores extends (infer S)[]
       : never
     : never
   : never
+
+export type WohooxPlugin = {
+  beforeInit?(initState, actions): { initState; actions }
+  onInit?(store: { name: string; state; actions }): void
+  onAdd?(storeName: string, value: any, keys: any[], target: any): void
+  onDelete?(storeName: string, keys: any[], target: any): void
+  onChange?(
+    storeName: string,
+    value: any,
+    keys: any[],
+    target: any,
+    oldValue: any,
+  ): void
+  onGet?(storeName: string, value: any, keys: any[], target: any): void
+}
